memory-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymilvus==2.2.8
redis==4.5.4
neo4j==5.6.0
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.24.0
pyarrow==11.0.0
numpy==1.24.2
sentence-transformers==2.2.2
prometheus-client==0.16.0
pytest==7.3.1
